[IMP] make it generic
moving mollie related code to a dedicated module
