+++ website_product_subscription_mollie_payment/__init__.py
@@ -1,0 +1,1 @@
+from . import models
