+++ website_product_subscription_mollie_payment/__openerp__.py
@@ -3,20 +3,19 @@
 #   Houssine Bakkali <houssine@coopiteasy.be>
 # License AGPL-3.0 or later (https://www.gnu.org/licenses/agpl.html).
 {
-    "name": "Website Product Subscription Online Payment",
+    "name": "Website Product Subscription Mollie Payment",
     "version": "9.0.1.0.1",
-    "depends": ["website_product_subscription",
-                "website_payment"],
+    "depends": ["website_product_subscription_online_payment",
+                "payment_mollie_official"],
     "author": "Coop IT Easy SCRL",
     "category": "Cooperative management",
     'website': "www.coopiteasy.be",
     "description": """
-        This module allows to pay the product subscription online
+        This module is the glue to allow product subscription online payment
+        "through mollie.
     """,
     'data': [
-        "views/online_payment_template.xml",
-        "views/subscription_request_view.xml",
-        "views/payment_acquirer_view.xml"
     ],
     'installable': True,
+    'auto_install': True,
 }
