+++ website_product_subscription_mollie_payment/models/__init__.py
@@ -1,0 +1,1 @@
+from . import mollie_payment
